// apps/web/types/jobRecommendations.type.ts
export interface JobRecommendation {
  yearsExperience: string;
  jobTitles: string[];
  roles: string[];
  responsibilities: string[];
  seniority: string[];
  industry: string[];
  hardSkills: string[];
  softSkills: string[];
  workAndTeamEnvironment: string[];
  companySizeFit: string[];
  careerTrack: string[];
}